Fixed broken test for connecting to new remote connection
It looks like this test has actually been broken for a while.  In the database setup file we had an api pointing to `localhost:8080/api` then tried to create a new remote pointing to the same uri which should fail but it was not. Since this is used several places, I made a new database xml file just for create.